remove _chapter index, go to first page directly
