+++ components/globe-component/globe-component.js
@@ -11,7 +11,8 @@ import Particles from './particles'
 
 const GLOBE_RADIUS = 5
 const WHITE = new THREE.Color(0xffffff)
-const vOffset = 15
+const tweenDuration = 1500
+const vOffset = 25
 const vOffsetFactor = vOffset / 100
 
 export default {
@@ -211,7 +212,7 @@ export default {
      */
     panAndZoom (from, to) {
       const tween = new Tween(from)
-        .to(to, 3000)
+        .to(to, tweenDuration)
         .on('update', ({ r, theta, phi }) => {
           const cart = polar2cartesian(r, theta, phi)
           this.camera.position.set(cart.x, cart.y, cart.z)
